requirements.txt
pyTelegramBotAPI==4.14.0
cairosvg
github_poster
prettymapp
google-generativeai

python-dotenv